+++ packages/core/build.config.ts
@@ -35,6 +35,7 @@ export default defineBuildConfig({
 		'src/plugins/generate/autoImport/index',
 		'src/plugins/generate/generatePages/index',
 		'src/plugins/generate/generateRouter/index',
+		'src/plugins/generate/generateUni/index',
 		'src/plugins/generate/generateVersion/index',
 		'src/plugins/guard/index',
 		'src/plugins/guard/envGuard/index',
@@ -52,7 +53,7 @@ export default defineBuildConfig({
 		'@': resolve(__dirname, './src')
 	},
 	replace: {
-		__PLUGIN_VERSION__: "\"1.2.0\""
+		__PLUGIN_VERSION__: "\"1.3.0\""
 	},
 	rollup: {
 		emitCJS: true,