+++ packages/uni-app/uni_modules/vite-plugin/js_sdk/plugins/generate/index.d.ts
@@ -1,8 +1,11 @@
 export { AutoImportOptions, BiomelintrcConfig, CacheConfig, DirConfig, DirConfigObject, DirsScanOptions, DtsConfig, DtsConfigObject, EslintrcConfig, ImportInline, ImportMeta, ImportsConfig, InlineImportConfig, PackagePresetOptions, PresetDefinition, Resolver, ScannedModule, TransformResult, VueDirectivesConfig, autoImport } from './autoImport/index.js';
 export { generateRouter } from './generateRouter/index.js';
 export { GenerateVersionOptions, OutputType, VersionFormat, VersionInfo, generateVersion } from './generateVersion/index.js';
-export { GeneratePagesOptions, RouteConfigBlock, SubPackageConfig, TabBarTemplate, generatePages } from './generatePages/index.js';
-export { G as GenerateRouterOptions, N as NameStrategy, O as OutputFormat, R as RouteConfig, a as RouteMeta, U as UniAppPageConfig, b as UniAppPagesJson, c as UniAppTabBarConfig } from '../../shared/vite-plugin.D2NWtthI.js';
+export { generatePages } from './generatePages/index.js';
+export { GenerateUniOptions, generateUni } from './generateUni/index.js';
+export { G as GeneratePagesOptions, R as RouteConfigBlock, S as SubPackageConfig, T as TabBarTemplate } from '../../shared/vite-plugin.CBG_QeBu.js';
+export { G as GenerateRouterOptions, N as NameStrategy, O as OutputFormat, R as RouteConfig, a as RouteMeta } from '../../shared/vite-plugin.B8yTanEe.js';
+export { U as UniAppPageConfig, a as UniAppPagesJson, b as UniAppTabBarConfig } from '../../shared/vite-plugin.CjZHnjC7.js';
 import '../../factory/index.js';
 import 'vite';
 import '../../shared/vite-plugin.BmNqGOzh.js';